requirements.txt
pyotp==2.9.0
sqlalchemy==2.0.37
fastapi==0.115.8
uvicorn==0.34.0
python-jose==3.3.0
bcrypt==4.2.1
redis==5.2.1
pymysql==1.1.1
requests==2.32.3
pyyaml==6.0.2
cryptography==44.0.0